+++ sourceCode/mainContainers/sceneContainers/customData_old.cpp
@@ -139,7 +139,7 @@ void CCustomData_old::serializeData(CSer &ar,const char* objectName,int scriptHa
             {
                 int s;
                 serObj[0] >> s;
-                for (size_t i=0;i<s;i++)
+                for (size_t i=0;i<size_t(s);i++)
                 {
                     int e;
                     int l;
